backend/src/main/java/com/amigoscode/customer/CustomerRowMapper.java
package com.amigoscode.customer;

import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;

import java.sql.ResultSet;
import java.sql.SQLException;

@Component
public class CustomerRowMapper implements RowMapper {
    @Override
    public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
        return new Customer(
                rs.getLong("id"),
                rs.getInt("age"),
                rs.getString("name"),
                rs.getString("email")
        );
    }
}
